[Develop]: login features.

+++ app/controllers/login_sessions_controller.rb
@@ -4,25 +4,27 @@ class LoginSessionsController < ApplicationController
   
   def create
     inputParams = loginParams
-    user = User.where(email: inputParams[:email]).first
-    if user && user.password == inputParams[:password]
-      session[:user_id] = user.id
-      session[:user_nickname] = user.nickname
-      session[:user_avatar_url] = user.avatar.url(:thumb)
+    
+    case inputParams[:type] when "sign in"
+      user = User.where(email: inputParams[:email]).first
+      if user && user.password == inputParams[:password]
+        session[:user_id] = user.id
+        session[:user_nickname] = user.nickname
+        session[:user_avatar_url] = user.avatar.url(:thumb)
+        session[:user_tier] = user.tier
+      end
+    when "sign out"
+      session.delete(:user_id)
+      session.delete(:user_nickname)
+      session.delete(:user_avatar_url)
+      session.delete(:user_tier)
     end
-  end
-
-
-  def destroy
-    session.delete(:user_id)
-    session.delete(:user_nickname)
-    session.delete(:user_avatar_url)
   end
   
 
 private
   def loginParams
-    params.require(:login_session).permit(:email, :password)
+    params.require(:login_session).permit(:email, :password, :type)
   end
   
 end
